refactor: My 컨트롤러 api 인터페이스 분리
